+++ services/web/server/src/simcore_service_webserver/rest/plugin.py
@@ -21,8 +21,8 @@ from .._meta import API_VTAG, api_version_prefix
 from ..security.plugin import setup_security
 from . import _handlers
 from .healthcheck import HealthCheck
-from .rest_settings import RestSettings, get_plugin_settings
 from .rest_utils import get_openapi_specs_path, load_openapi_specs
+from .settings import RestSettings, get_plugin_settings
 
 log = logging.getLogger(__name__)
