+++ services/web/server/src/simcore_service_webserver/rest/settings.py
